+++ src/codenarc-factory.js
@@ -5,12 +5,14 @@ const debug = require("debug")("npm-groovy-lint");
 const fse = require("fs-extra");
 const os = require("os");
 const xml2js = require("xml2js");
-const { npmGroovyLintRules } = require("./groovy-lint-rules.js");
+const { getNpmGroovyLintRules } = require("./groovy-lint-rules.js");
 const { evaluateRange, evaluateVariables, getSourceLines } = require("./utils.js");
 
 ////////////////////////////
 // Build codenarc options //
 ////////////////////////////
+
+const npmGroovyLintRules = getNpmGroovyLintRules();
 
 // Convert NPM-groovy-lint into codeNarc arguments
 // Create temporary files if necessary
@@ -38,7 +40,7 @@ async function prepareCodeNarcCall(options) {
     // Create ruleSet groovy file if necessary
     const ruleSetFileName = await manageCreateRuleSetFile(options);
     options.rulesets = ruleSetFileName;
-    if (ruleSetFileName.includes()) {
+    if (ruleSetFileName.includes("codeNarcTmpRs_")) {
         result.tmpRuleSetFileName = ruleSetFileName;
     }
 
@@ -190,7 +192,6 @@ async function manageCreateRuleSetFile(options) {
             return { ruleName: ruleNameShort };
         });
     }
-
     // Rules from config file, only if rulesets has not been sent as argument
     if (ruleSetsDef.length === 0 && options.rules) {
         for (const ruleName of Object.keys(options.rules)) {
@@ -234,7 +235,6 @@ async function manageCreateRuleSetFile(options) {
         debug(`CREATE RULESET tmp file ${tmpRuleSetFileName} generated from input options, as CodeNarc requires physical files`);
         return tmpRuleSetFileName;
     }
-    return "";
 }
 
 async function manageDeleteTmpFiles(tmpGroovyFileName, tmpRuleSetFileName) {