+++ src/config.js
@@ -7,10 +7,12 @@ const importFresh = require("import-fresh");
 const path = require("path");
 const stripComments = require("strip-json-comments");
 
+const defaultConfigFileName = ".groovylintrc-recommended.json";
+
 const configFilenames = [
+    ".groovylintrc.json",
     ".groovylintrc.js",
     ".groovylintrc.cjs",
-    ".groovylintrc.json",
     ".groovylintrc.yml",
     ".groovylintrc.yaml",
     ".groovylintrc",
@@ -19,6 +21,38 @@ const configFilenames = [
 
 // Load configuration from identified file, or find config file from a start path
 async function loadConfig(startPathOrFile) {
+    const configFilePath = await getConfigFileName(startPathOrFile);
+    // Load user configuration from file
+    let configUser = await loadConfigFromFile(configFilePath);
+    configUser.rules = await shortenRuleNames(configUser.rules || {});
+    // If config extends a standard one, merge it
+    configUser = await manageExtends(configUser);
+    return configUser;
+}
+
+// If extends defined, gather base level rules and append them to current rules
+async function manageExtends(configUser) {
+    if (configUser.extends) {
+        const baseConfigFilePath = await findConfigInPath(__dirname, [`.groovylintrc-${configUser.extends}.json`]);
+        let baseConfig = await loadConfigFromFile(baseConfigFilePath);
+        baseConfig.rules = await shortenRuleNames(baseConfig.rules || {});
+        // A config can extend another config that extends another config
+        baseConfig = await manageExtends(baseConfig);
+        // Delete doublons
+        for (const baseRuleName of Object.keys(baseConfig.rules)) {
+            for (const userRuleName of Object.keys(configUser.rules)) {
+                if (baseRuleName.includes(userRuleName)) {
+                    delete baseConfig.rules[baseRuleName];
+                }
+            }
+        }
+        configUser.rules = Object.assign(baseConfig.rules, configUser.rules);
+        delete configUser.extends;
+    }
+    return configUser;
+}
+
+async function getConfigFileName(startPathOrFile) {
     let configFilePath = null;
     const stat = await fse.lstat(startPathOrFile);
     // Find one of the config file formats at the root of the project or at upper directory levels
@@ -27,17 +61,9 @@ async function loadConfig(startPathOrFile) {
     }
     if (configFilePath == null) {
         // If not found, use .groovylintrc-recommended.js delivered with npm-groovy-lint
-        configFilePath = await findConfigInPath(__dirname, [".groovylintrc-recommended.js"]);
+        configFilePath = await findConfigInPath(__dirname, [defaultConfigFileName]);
     }
-    // Load user configuration from file
-    const configUser = loadConfigFromFile(configFilePath);
-    // If config extends a standard one, merge it
-    if (configUser.extends) {
-        const baseConfigFilePath = await findConfigInPath(__dirname, [`.groovylintrc-${configUser.extends}.js`]);
-        const baseConfig = loadConfigFromFile(baseConfigFilePath);
-        configUser.rules = Object.assign(baseConfig.rules, configUser.rules);
-    }
-    return configUser;
+    return configFilePath;
 }
 
 // try to  find a config file or config prop in package.json
@@ -47,7 +73,7 @@ async function findConfigInPath(directoryPath, configFilenamesIn) {
         if (await fse.exists(filePath)) {
             if (filename === "package.json") {
                 try {
-                    loadPackageJSONConfigFile(filePath);
+                    await loadPackageJSONConfigFile(filePath);
                     return filePath;
                 } catch (error) {
                     /* ignore */
@@ -65,27 +91,27 @@ async function findConfigInPath(directoryPath, configFilenamesIn) {
     return null;
 }
 
-function loadConfigFromFile(filePath) {
+async function loadConfigFromFile(filePath) {
     switch (path.extname(filePath)) {
         case ".js":
         case ".cjs":
-            return loadJSConfigFile(filePath);
+            return await loadJSConfigFile(filePath);
 
         case ".json":
             if (path.basename(filePath) === "package.json") {
-                return loadPackageJSONConfigFile(filePath);
+                return await loadPackageJSONConfigFile(filePath);
             }
-            return loadJSONConfigFile(filePath);
+            return await loadJSONConfigFile(filePath);
 
         case ".yaml":
         case ".yml":
-            return loadYAMLConfigFile(filePath);
+            return await loadYAMLConfigFile(filePath);
         default:
             return null;
     }
 }
 
-function loadJSConfigFile(filePath) {
+async function loadJSConfigFile(filePath) {
     debug(`Loading JS config file: ${filePath}`);
     try {
         return importFresh(filePath);
@@ -96,10 +122,10 @@ function loadJSConfigFile(filePath) {
     }
 }
 
-function loadJSONConfigFile(filePath) {
+async function loadJSONConfigFile(filePath) {
     debug(`Loading JSON config file: ${filePath}`);
     try {
-        return JSON.parse(stripComments(readFile(filePath)));
+        return JSON.parse(stripComments(await readFile(filePath)));
     } catch (e) {
         debug(`Error reading JSON file: ${filePath}`);
         e.message = `Cannot read config file: ${filePath}\nError: ${e.message}`;
@@ -112,7 +138,7 @@ function loadJSONConfigFile(filePath) {
     }
 }
 
-function loadYAMLConfigFile(filePath) {
+async function loadYAMLConfigFile(filePath) {
     debug(`Loading YAML config file: ${filePath}`);
 
     // lazy load YAML to improve performance when not used
@@ -120,7 +146,7 @@ function loadYAMLConfigFile(filePath) {
 
     try {
         // empty YAML file can be null, so always use
-        return yaml.safeLoad(readFile(filePath)) || {};
+        return yaml.safeLoad(await readFile(filePath)) || {};
     } catch (e) {
         debug(`Error reading YAML file: ${filePath}`);
         e.message = `Cannot read config file: ${filePath}\nError: ${e.message}`;
@@ -128,15 +154,14 @@ function loadYAMLConfigFile(filePath) {
     }
 }
 
-function loadPackageJSONConfigFile(filePath) {
+async function loadPackageJSONConfigFile(filePath) {
     debug(`Loading package.json config file: ${filePath}`);
     try {
-        const packageData = loadJSONConfigFile(filePath);
+        const packageData = await loadJSONConfigFile(filePath);
 
         if (!Object.hasOwnProperty.call(packageData, "groovylintConfig")) {
             throw Object.assign(new Error("package.json file doesn't have 'groovylintConfig' field."), { code: "GROOVYLINT_CONFIG_FIELD_NOT_FOUND" });
         }
-
         return packageData.groovylintConfig;
     } catch (e) {
         debug(`Error reading package.json file: ${filePath}`);
@@ -150,4 +175,14 @@ async function readFile(filePath) {
     return fileContent.replace(/^\ufeff/u, "");
 }
 
-module.exports = { loadConfig };
+// Remove rule category of rule name if defined. Ex: "basic.ConstantAssertExpression" beccomes "ConstantAssertExpression"
+async function shortenRuleNames(rules) {
+    const shortenedRules = {};
+    for (const ruleName of Object.keys(rules)) {
+        const ruleNameShort = ruleName.includes(".") ? ruleName.split(".")[1] : ruleName;
+        shortenedRules[ruleNameShort] = rules[ruleName];
+    }
+    return shortenedRules;
+}
+
+module.exports = { loadConfig, getConfigFileName };