+++ src/groovy-lint.js
@@ -8,7 +8,7 @@ const debug = require("debug")("npm-groovy-lint");
 const NpmGroovyLintFix = require("./groovy-lint-fix");
 const CodeNarcCaller = require("./codenarc-caller");
 const { prepareCodeNarcCall, parseCodeNarcResult, manageDeleteTmpFiles } = require("./codenarc-factory");
-const { loadConfig } = require("./config.js");
+const { loadConfig, getConfigFileName } = require("./config.js");
 const optionsDefinition = require("./options");
 const { computeStats, processOutput } = require("./output.js");
 
@@ -68,8 +68,10 @@ class NpmGroovyLint {
 
     // Call an existing NpmGroovyLint instance to request fix of errors
     async fixErrors(errorIds, optns = {}) {
-        // Run fixer on existing NpmGroovyLint instance
+        // Create and run fixer
         debug(`Fix errors for ${JSON.stringify(errorIds)} on existing NpmGroovyLint instance`);
+        const codeNarcFactoryResult = await prepareCodeNarcCall(this.options, this.jdeployRootPath);
+        this.setMethodResult(codeNarcFactoryResult);
         this.fixer = new NpmGroovyLintFix(this.lintResult, {
             verbose: optns.verbose || this.options.verbose,
             fixrules: optns.fixrules || this.options.fixrules,
@@ -78,13 +80,21 @@ class NpmGroovyLint {
         });
         await this.fixer.run({ errorIds: errorIds, propagate: true });
         this.lintResult = this.fixer.updatedLintResult;
-        // Control fix result by calling a new lint
-        await this.lintAgainAfterFix();
+        // Lint again after fix if requested (for the moment we prefer to trigger that from VsCode, for better UX)
+        if (optns.lintAgainAfterFix) {
+            // Control fix result by calling a new lint
+            await this.lintAgainAfterFix();
+        }
         // Compute stats & build output result
         this.lintResult = computeStats(this.lintResult);
         this.outputString = await processOutput(this.outputType, this.output, this.lintResult, this.options, this.fixer);
         // Delete Tmp file if existing
         manageDeleteTmpFiles(this.tmpGroovyFileName, this.tmpRuleSetFileName);
+    }
+
+    // Returns the full path of the configuration file
+    async getConfigFilePath() {
+        return getConfigFileName(this.options.config);
     }
 
     // Actions before call to CodeNarc
@@ -186,7 +196,7 @@ class NpmGroovyLint {
     // After CodeNarc call
     async postProcess() {
         // CodeNarc error
-        if (this.codeNarcStdErr && [null, "", undefined].includes(this.codeNarcStdOut)) {
+        if ((this.codeNarcStdErr && [null, "", undefined].includes(this.codeNarcStdOut)) || this.status > 0) {
             this.status = 1;
             console.error("GroovyLint: Error running CodeNarc: \n" + this.codeNarcStdErr);
         }